+++ search_result/index.js
@@ -1,68 +1,61 @@
 /// <reference types="ctx-core" />
 /// <reference types="./index.d.ts" />
-import { setter_computed_ } from '@ctx-core/nanostores'
 import { run } from 'ctx-core/function'
+import { memo_ } from 'ctx-core/rmemo'
 /** @typedef {Ctx} */
 /**
  * @param {search_result__params_T}params
  * @returns {search_store_T}
  */
-export function search_result$__new(params) {
+export function search_result__new(params) {
 	const { query_, data_, clear } = params
 	let { timeout } = params
 	let current_search_store
 	const search_store$ =
-		setter_computed_(query_, (
-			query,
-			_set
-		)=>{
-			if (!query) {
-				run(clear || (()=>{
+		memo_(search_store$=>{
+			if (!query_()) {
+				run(clear ?? (()=>{
 					set({
 						done: true,
 						loading: false,
-						query: query_.get(),
+						query: query_(),
 						data: []
 					})
 				}))
-				return
+				return search_store$.val
 			}
-			if (current_search_store?.query === query) {
-				return
+			if (current_search_store?.query === query_()) {
+				return search_store$.val
 			}
 			set({
 				done: false,
 				loading: true,
-				query: query_.get()
+				query: query_()
 			})
 			if (typeof typeof timeout !== 'number') timeout = 10000
 			Promise.race([
 				new Promise((_res, rej)=>{
-					setTimeout(()=>rej(new Error('Timeout'))
-						, timeout)
+					setTimeout(()=>
+						rej(new Error('Timeout')), timeout)
 				}),
 				run(async ()=>{
-					const data = await data_({ query })
-					const $query_ = query_.get()
-					if (query === $query_) {
+					const data = await data_({ query: query_() })
+					const $query_ = query_()
+					if (query_() === $query_) {
 						set({
 							done: true,
 							loading: false,
-							query,
+							query: query_(),
 							data
 						})
 					}
 				})
 			]).then()
+			return search_store$.val
 			function set(search_store) {
 				current_search_store = search_store
-				_set(search_store)
+				search_store$._ = search_store
 			}
 		})
 	return search_store$
-}
-export {
-	search_result$__new as search_store__,
-	search_result$__new as search_result_store_,
-	search_result$__new as _store__search_result
 }