+++ search_result/index.test.ts
@@ -1,9 +1,9 @@
-import { atom } from '@ctx-core/nanostores'
-import { install, InstalledClock } from '@sinonjs/fake-timers'
+import { install, type InstalledClock } from '@sinonjs/fake-timers'
 import { sleep } from 'ctx-core/function'
+import { sig_ } from 'ctx-core/rmemo'
 import { test } from 'uvu'
 import { equal } from 'uvu/assert'
-import { search_result$__new } from '../index.js'
+import { search_result__new } from '../index.js'
 let clock:InstalledClock
 test.before(()=>{
 	clock = install()
@@ -12,32 +12,31 @@ test.after(()=>{
 	clock.uninstall()
 })
 test('search_result$__new|query_: blank -> present|single run', async ()=>{
-	const query_ = atom('')
+	const query_ = sig_('')
 	const data_ = async ()=>{
 		await sleep(10)
 		return [0, 1, 2]
 	}
-	const search_store$ = search_result$__new({
-		query_,
-		data_
-	})
-	equal(search_store$.$, {
+	const search_store$ =
+		search_result__new({
+			query_,
+			data_
+		})
+	equal(search_store$(), {
 		done: true,
 		loading: false,
 		query: '',
 		data: []
 	})
-	query_.set('query-value')
-	await clock.tickAsync(11)
-	await clock.runAllAsync()
-	equal(search_store$.$, {
+	query_._ = 'query-value'
+	equal(search_store$(), {
 		done: false,
 		loading: true,
 		query: 'query-value',
 	})
 	await clock.tickAsync(11)
 	await clock.runAllAsync()
-	equal(search_store$.$, {
+	equal(search_store$(), {
 		done: true,
 		loading: false,
 		query: 'query-value',